Fix strict type error in EntryQueryOptions::fromRequest()
Cast $request->take to int before passing to limit(). HTTP request
values are always strings, which fails under declare(strict_types=1).

+++ src/telescope/src/Storage/EntryQueryOptions.php
@@ -49,7 +49,7 @@ class EntryQueryOptions
             ->beforeSequence($request->before) // @phpstan-ignore-line
             ->tag($request->tag) // @phpstan-ignore-line
             ->familyHash($request->family_hash) // @phpstan-ignore-line
-            ->limit($request->take ?? 50);
+            ->limit((int) ($request->take ?? 50));
     }
 
     /**
